fix(fetch): honor the timeout a registry fetcher was created with
Review follow-up. `createFetchFromRegistry` fell back to the default
timeout whenever a request carried none of its own, so the `fetchTimeout`
the install layer passes as a default reached only the dispatcher's
connect bound. Now that the same value drives the header and body
bounds, ignoring it would leave `fetch-timeout` without effect on the
requests that matter.

A `timeout` of `0` now also disables the SOCKS handshake bound, which
the socks library expresses as a delay rather than an off switch.

The test server both suites trickle through moves to `test/utils`.
