+++ pnpm11/network/fetch/test/fetch.test.ts
@@ -1,11 +1,10 @@
 /// <reference path="../../../__typings__/index.d.ts"/>
-import { createServer, type ServerResponse } from 'node:http'
-import type { AddressInfo } from 'node:net'
-
 import { afterEach, describe, expect, jest, test } from '@jest/globals'
 import { requestRetryLogger } from '@pnpm/core-loggers'
 import { clearDispatcherCache, fetch } from '@pnpm/network.fetch'
 import { type Dispatcher, getGlobalDispatcher, MockAgent, setGlobalDispatcher } from 'undici'
+
+import { startServer } from './utils/trickleServer.js'
 
 test('metadata retry logs redact signed URL parameters', async () => {
   const originalDispatcher = getGlobalDispatcher()
@@ -124,28 +123,3 @@ describe('the fetch timeout measures inactivity, not total time', () => {
   })
 })
 
-interface TestServer extends AsyncDisposable {
-  url: string
-}
-
-async function startServer (respond: (res: ServerResponse) => void): Promise<TestServer> {
-  const server = createServer((_req, res) => {
-    res.writeHead(200, { 'content-type': 'application/octet-stream' })
-    respond(res)
-  })
-  await new Promise<void>((resolve) => {
-    server.listen(0, '127.0.0.1', resolve)
-  })
-  const { port } = server.address() as AddressInfo
-  return {
-    url: `http://127.0.0.1:${port}/`,
-    async [Symbol.asyncDispose] () {
-      server.closeAllConnections()
-      await new Promise<void>((resolve) => {
-        server.close(() => {
-          resolve()
-        })
-      })
-    },
-  }
-}
